dnssec-checker-master/requirements.txt
dnspython3>=1.15.0
validate_email>=1.3
